* Use gcc13 -fanalyzer and clang-devel scan-build (static analyzers) * The following changes were directed by the static analyzer * Check/use -fno-delete-null-pointer-checks in configure * Fix pop_integer() to return correct type * Use fseek() instead of rewind() so clang doesn't complain about errno * Don't check a pointer after using it so as to not offend gcc13 * If you pass no args to $unshift() or $push(), get nothing back * Fix uninitialized clue in $currchans() [this was an actual bug] * Strip down $globi() so I can fix static analysis complaints * Fix uninitialized variable in $hookctl() [this was an actual bug] * Don't just assume open() doesn't fail in opento() [actual bug] * ifdef out split_message() and friends because i don't understand     what gcc13 is complaining about there and I don't use it * Oh, you get the idea.  more of the same as above * Add /window lister "refnums" (see UPDATES)

+++ source/expr2.c
@@ -941,7 +941,7 @@ __inline static	TOKEN	pop_token (expr_info *c)
 		return c->stack[c->sp--];
 }
 
-__inline static	double		pop_integer (expr_info *c)
+__inline static	intmax_t		pop_integer (expr_info *c)
 {
 	return get_token_integer(c, pop_token(c));
 }
